@@ -11,12 +11,11 @@
   },
   "dependencies": {
     "axios": "^0.21.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "cross-env": "^7.0.3"
+    "cross-env": "^7.0.3",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-scripts": "4.0.3"
   },
-  "devDependencies": {},
   "author": "",
   "license": "ISC",
   "browserslist": {
