@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import UserRegistration from './components/UserRegistration';
 
 const App = () => (
@@ -8,4 +8,6 @@ const App = () => (
     </div>
 );
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(<App />);
